Moved to CDN Model
Images and other resources have been moved to cdn.reformationfundraising.com in order to abstract the problem of their organisation outside of this repository.

I am aware that there is likely a better way of potentially forking this using git, but given that I am the only one maintaining the repository I don't see a need to make things more complicated for myself.